requirements.txt
mysql-connector-python
pyspark
python-dotenv